Move files according to dependencies
