+++ app/src/main/java/org/researchstack/sampleapp/SampleDataProvider.java
@@ -3,7 +3,7 @@ import android.content.Context;
 
 import org.researchstack.backbone.ResourcePathManager;
 import org.researchstack.backbone.result.TaskResult;
-import org.researchstack.sampleapp.bridge.BridgeDataProvider;
+import org.sagebionetworks.bridge.researchstack.BridgeDataProvider;
 import org.researchstack.skin.ResourceManager;
