+++ app/src/main/java/org/researchstack/sampleapp/SampleResearchStack.java
@@ -12,7 +12,7 @@ import org.researchstack.backbone.storage.file.FileAccess;
 import org.researchstack.backbone.storage.file.PinCodeConfig;
 import org.researchstack.backbone.storage.file.SimpleFileAccess;
 import org.researchstack.backbone.storage.file.aes.AesProvider;
-import org.researchstack.sampleapp.bridge.BridgeEncryptedDatabase;
+import org.sagebionetworks.bridge.researchstack.BridgeEncryptedDatabase;
 import org.researchstack.skin.AppPrefs;
 import org.researchstack.skin.DataProvider;
 import org.researchstack.skin.PermissionRequestManager;
