--- app/src/main/java/org/researchstack/sampleapp/bridge/IdentifierHolder.java
@@ -1,7 +1,0 @@
-package org.researchstack.sampleapp.bridge;
-
-public class IdentifierHolder
-{
-    public String type;
-    public String identifier;
-}
