+++ app/src/main/java/org/sagebionetworks/bridge/researchstack/BridgeEncryptedDatabase.java
@@ -1,4 +1,5 @@
-package org.researchstack.sampleapp.bridge;
+package org.sagebionetworks.bridge.researchstack;
+
 import android.content.Context;
 
 import net.sqlcipher.database.SQLiteDatabase;
@@ -6,6 +7,8 @@ import net.sqlcipher.database.SQLiteDatabase;
 import org.researchstack.backbone.storage.database.sqlite.SqlCipherDatabaseHelper;
 import org.researchstack.backbone.storage.database.sqlite.UpdatablePassphraseProvider;
 import org.researchstack.backbone.utils.LogExt;
+import org.sagebionetworks.bridge.android.upload.UploadQueue;
+import org.sagebionetworks.bridge.researchstack.upload.UploadRequest;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -14,71 +17,55 @@ import co.touchlab.squeaky.db.sqlcipher.SQLiteDatabaseImpl;
 import co.touchlab.squeaky.table.TableUtils;
 
 
-public class BridgeEncryptedDatabase extends SqlCipherDatabaseHelper implements UploadQueue
-{
-    public BridgeEncryptedDatabase(Context context, String name, SQLiteDatabase.CursorFactory cursorFactory, int version, UpdatablePassphraseProvider passphraseProvider)
-    {
+public class BridgeEncryptedDatabase extends SqlCipherDatabaseHelper implements UploadQueue {
+    public BridgeEncryptedDatabase(Context context, String name, SQLiteDatabase.CursorFactory cursorFactory, int version, UpdatablePassphraseProvider passphraseProvider) {
         super(context, name, cursorFactory, version, passphraseProvider);
     }
 
     @Override
-    public void onCreate(SQLiteDatabase sqLiteDatabase)
-    {
+    public void onCreate(SQLiteDatabase sqLiteDatabase) {
         super.onCreate(sqLiteDatabase);
-        try
-        {
+        try {
             TableUtils.createTables(new SQLiteDatabaseImpl(sqLiteDatabase), UploadRequest.class);
-        }
-        catch(SQLException e)
-        {
+        } catch (SQLException e) {
             throw new RuntimeException(e);
         }
     }
 
     @Override
-    public void onUpgrade(SQLiteDatabase sqLiteDatabase, int oldVersion, int newVersion)
-    {
+    public void onUpgrade(SQLiteDatabase sqLiteDatabase, int oldVersion, int newVersion) {
         super.onUpgrade(sqLiteDatabase, oldVersion, newVersion);
         // handle future db upgrades here
     }
 
-    public void saveUploadRequest(UploadRequest uploadRequest)
-    {
+    @Override
+    public void saveUploadRequest(UploadRequest uploadRequest) {
         LogExt.d(this.getClass(), "saveUploadRequest() id: " + uploadRequest.id);
 
-        try
-        {
+        try {
             this.getDao(UploadRequest.class).createOrUpdate(uploadRequest);
-        }
-        catch(SQLException e)
-        {
+        } catch (SQLException e) {
             throw new RuntimeException(e);
         }
     }
 
-    public List<UploadRequest> loadUploadRequests()
-    {
-        try
-        {
+    @Override
+    public List<UploadRequest> loadUploadRequests() {
+        try {
             return this.getDao(UploadRequest.class).queryForAll().orderBy("id DESC").list();
-        }
-        catch(SQLException e)
-        {
+        } catch (SQLException e) {
             throw new RuntimeException(e);
         }
     }
 
-    public void deleteUploadRequest(UploadRequest request)
-    {
+    @Override
+    public void deleteUploadRequest(UploadRequest request) {
 
         LogExt.d(this.getClass(), "deleteUploadRequest() id: " + request.id);
 
-        try
-        {
+        try {
             this.getDao(UploadRequest.class).delete(request);
-        }
-        catch(SQLException e)
-        {
+        } catch (SQLException e) {
             throw new RuntimeException(e);
         }
     }
