+++ app/src/main/java/org/sagebionetworks/bridge/sdk/rest/model/SignUpBody.java
@@ -1,4 +1,4 @@
-package org.researchstack.sampleapp.bridge.body;
+package org.sagebionetworks.bridge.sdk.rest.model;
 public class SignUpBody
 {
