+++ gradle.properties
@@ -17,3 +17,5 @@
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
 
+studyId="rs-sample-app"
+studyName="ResearchStackSampleApp"
